[TASK] Create localize method for multiple usage
Create new generateTranslatedRecords function for multiple
usage in different tablehandlers.
Also extract the localize and copy functions.
