+++ Classes/TcaDataGenerator/TableHandler/General.php
@@ -66,71 +66,6 @@ class General extends AbstractTableHandler implements TableHandlerInterface
             [ 'uid' => $fieldValues['uid'] ]
         );
 
-        $demoLanguages = $recordFinder->findUidsOfDemoLanguages();
-
-        $translatedRecord = -42;
-        foreach ($demoLanguages as $demoLanguageIndex => $demoLanguageUid) {
-            switch ($demoLanguageIndex) {
-                case 0:
-                    $this->copyRecordToLanguage($tableName, $fieldValues['uid'], $demoLanguageUid);
-                    break;
-                case 1:
-                    $result = $this->localizeRecord($tableName, $fieldValues['uid'], $demoLanguageUid);
-                    $translatedRecord = $result[$tableName][$fieldValues['uid']];
-                    break;
-                case 2:
-                    $result = $this->localizeRecord($tableName, $translatedRecord, $demoLanguageUid);
-                    $translatedRecord = $result[$tableName][$translatedRecord];
-                    break;
-                case 3:
-                    $this->copyRecordToLanguage($tableName, $translatedRecord, $demoLanguageUid);
-                    break;
-                default:
-                    $this->localizeRecord($tableName, $fieldValues['uid'], $demoLanguageUid);
-                    break;
-            }
-        }
-    }
-
-    /**
-     * @param string $tableName
-     * @param int $uid
-     * @param int $languageId
-     * @return array
-     */
-    protected function localizeRecord($tableName, $uid, $languageId)
-    {
-        $commandMap = [
-            $tableName => [
-                $uid => [
-                    'localize' => $languageId,
-                ],
-            ],
-        ];
-        $dataHandler = GeneralUtility::makeInstance(DataHandler::class);
-        $dataHandler->start([], $commandMap);
-        $dataHandler->process_cmdmap();
-        return $dataHandler->copyMappingArray;
-    }
-
-    /**
-     * @param string $tableName
-     * @param int $uid
-     * @param int $languageId
-     * @return array
-     */
-    protected function copyRecordToLanguage($tableName, $uid, $languageId)
-    {
-        $commandMap = [
-            $tableName => [
-                $uid => [
-                    'copyToLanguage' => $languageId,
-                ],
-            ],
-        ];
-        $dataHandler = GeneralUtility::makeInstance(DataHandler::class);
-        $dataHandler->start([], $commandMap);
-        $dataHandler->process_cmdmap();
-        return $dataHandler->copyMappingArray;
+        $this->generateTranslatedRecords($tableName, $fieldValues);
     }
 }
